pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "pfs-ics-utils"
description = "Common util tools for the Subaru ICS actors"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "astropy",
    "astropy_iers_data",
    "fitsio",
    "gitpython",
    "numpy",
    "pandas",
    "psycopg2-binary",
    "pytz",
    "scipy",
    "sdss-actorcore",
    "sdss-opscore",
    "sqlalchemy",
    "twisted",
]
dynamic = ["version"]

[tool.setuptools.packages.find]
where = ["python/"]
include = ["ics", "ics.*"]

[tool.black]
line-length = 110
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 110
known_first_party = ["ics"]

[tool.lsst_versions]
write_to = "python/ics/utils/version.py"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"  # Recommended as best practice

[tool.pydocstyle]
convention = "numpy"
